16. GROQ+OLLAMA Chatbot/pyproject.toml
[project]
name = "openai-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.1",
    "langchain-openai>=1.1.7",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.1",
]
